+++ teacher.rb
@@ -1,4 +1,4 @@
-require_relative 'person_class'
+require_relative 'person'
 
 class Teacher < Person
   def initialize(specialization, age, name = 'Unknown', parent_permission: true)
